refactor(RoomView): split RightButtons into dispatcher and per-context components
