+++ app/views/RoomView/components/RightButtons/RightButtons.tsx
@@ -1,0 +1,40 @@
+import { type ReactElement } from 'react';
+import { useStore } from 'zustand';
+
+import { type RoomStore } from '../../definitions';
+import { OmnichannelRightButtons } from './OmnichannelRightButtons';
+import { RoomRightButtons } from './RoomRightButtons';
+import { ThreadRightButtons } from './ThreadRightButtons';
+
+interface IRightButtonsProps {
+	rid?: string;
+	tmid?: string;
+	roomStore: RoomStore;
+}
+
+const RightButtons = ({ rid, tmid, roomStore }: IRightButtonsProps): ReactElement | null => {
+	const room = useStore(roomStore, s => s.room);
+
+	if (!rid) {
+		return null;
+	}
+
+	if (room.status === 'INVITED') {
+		return null;
+	}
+
+	if (room.t === 'l') {
+		if (room.status === 'queued') {
+			return null;
+		}
+		return <OmnichannelRightButtons rid={rid} roomStore={roomStore} />;
+	}
+
+	if (tmid) {
+		return <ThreadRightButtons tmid={tmid} />;
+	}
+
+	return <RoomRightButtons rid={rid} roomStore={roomStore} />;
+};
+
+export default RightButtons;
